requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
art==6.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
numpy==1.26.4
openai==1.21.0
py==1.11.0
PyAudio==0.2.13
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
pynput==1.7.6
pyobjc-core==10.2
pyobjc-framework-ApplicationServices==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-CoreText==10.2
pyobjc-framework-Quartz==10.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.4.16
requests==2.31.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.1
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
